+++ ModelViewer/ViewModels/ConnectPageViewModel.cpp
@@ -1,8 +1,12 @@
 #include "pch.h"
 #include "ConnectPageViewModel.h"
 #include "DelegateCommand.h"
+#include "LoadingWrapper.h"
+#include <pplawait.h>
+#include <chrono>
 
 using namespace ViewModels;
+using namespace std::chrono;
 
 ConnectPageViewModel::ConnectPageViewModel()
 {
@@ -11,6 +15,18 @@ ConnectPageViewModel::ConnectPageViewModel()
 
 void ConnectPageViewModel::ExecuteConnectCommand(Object^ param)
 {
+}
+
+future<void> ConnectPageViewModel::Connect()
+{
+	Utility::Out(L"At Start of Connect");
+
+	// RAII-style for ensuring that the progress gets cleared robustly
+	auto loader = make_unique<LoadingWrapper>([this]() { Loading = true; }, [this]() { Loading = false; });
+
+	//co_await 5s;
+
+	Utility::Out(L"Connected");
 }
 
 bool ConnectPageViewModel::Loading::get()
